#699  Review #1. Move CatalogContext to context module
